requirements.txt
Django==4.1.4
psycopg2-binary==2.9.5
django-extensions==3.2.1
django-ckeditor==6.5.1
sorl-thumbnail==12.9.0
Pillow==9.3.0
